order store refactor (initial)
