+++ core/modules-entry.ts
@@ -1,14 +1,12 @@
 
 import { VueStorefrontModule } from '@vue-storefront/core/lib/module'
 import { Cms } from './modules/cms'
-import { Order } from './modules/order'
 import { User } from './modules/user'
 import { registerModules } from 'src/modules'
 
 // @deprecated from 2.0,  use registerModule instead
 export const enabledModules: VueStorefrontModule[] = [
   Cms,
-  Order,
   User,
   ...registerModules
 ]
