+++ core/modules/order/store/index.ts
@@ -5,7 +5,7 @@ import getters from './getters'
 import RootState from '@vue-storefront/core/types/RootState'
 import OrderState from '../types/OrderState'
 
-export const module: Module<OrderState, RootState> = {
+export const orderStore: Module<OrderState, RootState> = {
   namespaced: true,
   state: {
     last_order_confirmation: null,
